Fixed the warnings

+++ src/process_bpmn.rs
@@ -1,3 +1,4 @@
+#[derive(Serialize, Deserialize, Debug)]
 pub enum Process {
-    ProcessId(usize),
+    ProcessId(i128),
 }
